dont create empty repo if one exists (fix prev commit)

+++ update_repository.sh
@@ -53,10 +53,14 @@ fi
 chown -R builder:builder /home/builder/workspace
 
 # create an empty repository file
-sudo --user builder \
-    tar cvfz /home/builder/workspace/$INPUT_REPONAME.db.tar.gz -T /dev/null
+if [ ! -f /home/builder/workspace/$INPUT_REPONAME.db.tar.gz ] ; then
+    sudo --user builder \
+        tar cvfz /home/builder/workspace/$INPUT_REPONAME.db.tar.gz -T /dev/null
+fi
+if [ ! -f /home/builder/workspace/$INPUT_REPONAME.db ] ; then
 sudo --user builder \
     ln -s $INPUT_REPONAME.db.tar.gz /home/builder/workspace/$INPUT_REPONAME.db
+fi
 
 ## Register the local repository with pacman.
 cat >> /etc/pacman.conf <<-EOF
